requirements.txt
openai
pyautogui
pyperclip
python-dotenv